Split the image API and sqlite functions
